+++ src/plx_core.c
@@ -36,6 +36,7 @@ extern const PlxDialect plx_php_dialect;
 extern const PlxDialect plx_js_dialect;
 extern const PlxDialect plx_py_dialect;
 extern const PlxDialect plx_cobol_dialect;
+extern const PlxDialect plx_plsql_dialect;
 
 /* ---- dialect registry ------------------------------------------------- */
 
@@ -203,4 +204,5 @@ _PG_init(void)
 	plx_register_dialect(&plx_js_dialect);
 	plx_register_dialect(&plx_py_dialect);
 	plx_register_dialect(&plx_cobol_dialect);
+	plx_register_dialect(&plx_plsql_dialect);
 }